requirements.txt
python-dotenv==1.0.1
asyncua==1.1.0
matplotlib==3.9.1
